+++ includes/class-tmsm-woocommerce-booking-thalasso.php
@@ -110,6 +110,7 @@ class Tmsm_Woocommerce_Booking_Thalasso {
 		if ( function_exists( '_is_elementor_installed' ) ) {
 			require_once plugin_dir_path( dirname( __FILE__ ) ) . 'includes/elementor/elementor-tag-bookroombuttonlabel.php';
 			require_once plugin_dir_path( dirname( __FILE__ ) ) . 'includes/elementor/elementor-tag-accommodationurl.php';
+			require_once plugin_dir_path( dirname( __FILE__ ) ) . 'includes/elementor/elementor-tag-packageratestable.php';
 		}
 
 		/**